g2tester/requirements.txt
pygame
termcolor
smbus2
sparkfun-qwiic
sparkfun-top-phat-button
adafruit-circuitpython-neopixel
RPi.GPIO
testproject-python-sdk
rpi_ws281x